requirements.txt
django==5.2.12
djangorestframework==3.15.2
django-filter==25.2
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
